pyproject.toml
[project]
name = "assistant4"
version = "4.0.0"
description = "AI agent 4"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "croniter>=1.4.0",
    "openai>=1.0.0",
    "openai-agents[sqlalchemy,mcp]>=0.2.9",
    "sqlite-vec>=0.1.0",
    "aiogram>=3.24.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.20.0",
    "humanize>=4.9.0",
    "questionary>=2.0.0",
    "keyring>=25.0.0",
    "ddgs>=9.0.0",
    "tavily-python>=0.5.0",
    "aioimaplib>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "ag-ui-protocol>=0.1.13",
]

[project.optional-dependencies]
litellm = [
    "openai-agents[litellm]",
]
observability = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.11.0",
    "types-PyYAML>=6.0.12.20250915",
    "types-croniter>=6.0.0.20250809",
    "import-linter>=2.0.0",
    "bandit>=1.7.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "vulture>=2.14",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "sandbox/extensions"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]  # flake8 + isort + bugbear + pyupgrade
ignore = ["E402", "E501"]

[tool.ruff.format]
# drop-in replacement for black

[tool.importlinter]
root_package = "core"
include_external_packages = true
# Extensions are loaded dynamically via package imports (sandbox.extensions.*);
# this contract only enforces one-way core -> extensions isolation.

[[tool.importlinter.contracts]]
name = "Core is independent of extensions"
type = "forbidden"
source_modules = ["core"]
forbidden_modules = ["sandbox", "extensions"]

[tool.mypy]
python_version = "3.12"
strict = true
# Optional: pyright (Microsoft) is faster and handles Python 3.12+ type hints well; can complement mypy.
warn_unused_configs = true
warn_unreachable = true
warn_no_return = true
show_error_codes = true
exclude = "(^|/)(\\.venv|tests|scripts|sandbox/extensions)(/|$)"

[[tool.mypy.overrides]]
module = ["tests.*", "scripts.*", "migrations.*"]
disallow_untyped_defs = false
check_untyped_defs = false

[[tool.mypy.overrides]]
module = ["app.web.*", "app.di.*"]
disallow_untyped_decorators = false

[tool.bandit]
exclude_dirs = ["tests", "test_*.py"]
skips = ["B101"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.vulture]
min_confidence = 80
paths = [
    "core",
    "sandbox/extensions",
    "onboarding",
    "supervisor",
    "scripts/vulture_allowlist.py",
]
exclude = [".venv", "tests", "**/__pycache__"]

[dependency-groups]
dev = [
    "honcho>=2.0.0",
]
